Extract SettingsFragment into its own file.
